requirements.txt
asgiref==3.9.1
Django==5.2.5
Faker==37.6.0
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2

clickhouse-connect==0.8.18
clickhouse-driver==0.2.9